Database Fix
Modifies the return type of the DAO methods 'setCoverPhotoId' and 'deleteTripById' to Int., Update and Delete must return an Int based on room documentation

+++ app/src/main/java/com/aptivist/tripsupiicsaapp/data/local/database/daos/TripDao.kt
@@ -31,8 +31,8 @@ interface TripDao {
     suspend fun upsert(trip: TripEntity): Long
 
     @Query("UPDATE trips Set cover_photo_id = :photoId, modified_at = :modifiedAt WHERE id = :tripId")
-    suspend fun setCoverPhotoId(tripId: Long, photoId: Long?, modifiedAt: Date = Date()) : Long
+    suspend fun setCoverPhotoId(tripId: Long, photoId: Long?, modifiedAt: Date = Date()) : Int
 
     @Query("DELETE FROM trips WHERE id = :tripId")
-    suspend fun deleteTripById(tripId: Long) : Long
+    suspend fun deleteTripById(tripId: Long) : Int
 }
